set up erros lib and errorHandler and implement in auth and character controllers

+++ controllers/character.js
@@ -1,24 +1,29 @@
 import express from 'express'
 import Character from '../models/character.js'
 import secureRoute from '../middleware/secureRoute.js'
+import { NotFound, AlreadyExists, Unauthorised } from '../lib/errors.js'
 
 const router = express.Router()
 
 router.get('/characters', async function characterIndex(req, res, next) {
   try {
+    throw new NotFound()
     const characters = await Character.find();
+    
+    if (!characters) throw new NotFound()
     return res.status(200).json(characters);
   } catch (err) {
-    console.log(err)
+      next(err)
   }
 })
 
 router.get('/characters/:characterId', async function characterShow(req, res, next) {
   try {
     const character = await Character.findById(req.params.characterId);
+    if (!character) throw new NotFound()
     return res.status(200).json(character);
   } catch (err) {
-    console.log(err)
+    next(err)
   }
 })
 
@@ -26,15 +31,14 @@ router.post('/characters', secureRoute, async function characterNew(req, res, ne
   try {
     if (res.locals.currentUser.isAdmin) {
       const existingCharacter = await Character.findOne({ name: req.body.name });
-      if (existingCharacter) throw new Error('Already exists');
+      if (existingCharacter) throw new AlreadyExists()
       const newCharacter = await Character.create(req.body);
       return res.status(201).json(newCharacter);
-      
+
     } else {
-      throw new Error('Unauthorised');
+      throw new Unauthorised()
     }
   } catch (err) {
-    console.log(err)
     next(err)
   }
 })
@@ -43,15 +47,15 @@ router.put('/characters/:characterId', secureRoute, async function characterEdit
   try {
     if (res.locals.currentUser.isAdmin) {
       const characterToUpdate = await Character.findById(req.params.characterId);
+      if (!characterToUpdate) throw new NotFound()
       Object.assign(characterToUpdate, req.body);
       await characterToUpdate.save();
       return res.status(202).json(characterToUpdate);
 
     } else {
-      throw new Error('Unauthorised')
+      throw new Unauthorised()
     }
   } catch (err) {
-    console.log(err)
     next(err)
   }
 })
@@ -60,14 +64,14 @@ router.delete('/characters/:characterId', secureRoute, async function characterD
   try {
     if (res.locals.currentUser.isAdmin) {
       const characterToDelete = await Character.findById(req.params.characterId);
+      if (!characterToDelete) throw new NotFound()
       await characterToDelete.deleteOne();
       return res.sendStatus(204);
-    
+
     } else {
-      throw new Error('Unauthorised');
+      throw new Unauthorised()
     }
   } catch (err) {
-    console.log(err)
     next(err)
   }
 })
